+++ R/to_value.R
@@ -17,7 +17,8 @@
 #'          level (if factor is \code{\link{numeric}}) or to \code{1} (if factor levels
 #'          are not numeric).
 #' @param keep.labels logical, if \code{TRUE}, former factor levels will be added as
-#'          value labels. See \code{\link{set_labels}} for more details.
+#'          value labels. For numeric factor levels, values labels will be used,
+#'          if present. See 'Examples' and \code{\link{set_labels}} for more details.
 #' @return A numeric variable with values ranging either from \code{start.at} to
 #'           \code{start.at} + length of factor levels, or to the corresponding
 #'           factor levels (if these were numeric). Or a data frame with numeric
@@ -31,8 +32,7 @@
 #' table(to_value(test))
 #' hist(to_value(test, 0))
 #'
-#' # set lowest value of new variable
-#' # to "5".
+#' # set lowest value of new variable to "5".
 #' table(to_value(test, 5))
 #'
 #' # numeric factor keeps values
@@ -48,6 +48,20 @@
 #' # starting at 1
 #' dummy <- factor(c("D", "F", "H"))
 #' table(to_value(dummy))
+#'
+#' # for numeric factor levels, value labels will be used, if present
+#' dummy1 <- factor(c("3", "4", "6"))
+#' set_labels(dummy1) <- c("first", "2nd", "3rd")
+#' dummy1
+#' to_value(dummy1)
+#'
+#' # for non-numeric factor levels, these will be used.
+#' # value labels will be ignored
+#' dummy2 <- factor(c("D", "F", "H"))
+#' set_labels(dummy2) <- c("first", "2nd", "3rd")
+#' dummy2
+#' to_value(dummy2)
+#'
 #'
 #' @export
 to_value <- function(x, start.at = NULL, keep.labels = TRUE) {
@@ -73,10 +87,10 @@ to_value_helper <- function(x, start.at, keep.labels) {
   labels <- NULL
   # is already numeric?
   if (is.numeric(x)) return(x)
+  # get labels
+  labels <- get_labels(x, attr.only = T, include.values = "n")
   # is character?
   if (is.character(x)) {
-    # get labels
-    labels <- get_labels(x, attr.only = T, include.values = "n")
     # has labels?
     if (!is.null(labels)) {
       # sort labels correctly
@@ -86,10 +100,10 @@ to_value_helper <- function(x, start.at, keep.labels) {
     # convert to factor
     x <- as.factor(x)
   }
-  # retrieve "value labels"
-  if (is.null(labels)) labels <- levels(x)
   # check if we have numeric factor levels
   if (is_num_fac(x)) {
+    # retrieve "value labels"
+    if (is.null(labels)) labels <- levels(x)
     # convert to numeric via as.vector
     new_value <- as.numeric(as.vector((x)))
     # new minimum value?
@@ -101,6 +115,8 @@ to_value_helper <- function(x, start.at, keep.labels) {
       new_value <- new_value + val_diff
     }
   } else {
+    # use non-numeric factor levels as new labels
+    labels <- levels(x)
     # check start.at value
     if (is.null(start.at)) start.at <- 1
     # get amount of categories