+++ NEWS.md
@@ -22,6 +22,7 @@
 * `to_label` gets a `var.label` and `drop.levels` argument, and now preserves variable labels by default.
 * Argument `def.value` in `get_label` now also applies to data frame arguments.
 * If factor levels are numeric and factor has value labels, these are used in `to_value` by default.
+* `to_factor` no longer generates `NA` or `NaN`-levels when converting input into factors.
 
 ## Bug fixes